+++ CHANGELOG.md
@@ -8,6 +8,7 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 ## Version 0.10.3 - Unreleased
 
 ### Added
+* Added `ubelt.Path`, as an extension and quicker-to-type version of pathlib.Path with extra functionality.
 * Added `progkw` as argument to `JobPool.as_completed` to control progress reporting
 * Added `progkw` as argument to `ub.download` / `ub.grabdat` to control progress reporting
 * Added `util_zip` with the `zopen` function. Access a file inside a zipfile with a standard `open` like interface.